Move noop out of common to avoid circular dependency
