+++ src/dynamic-import-csp.js
@@ -1,5 +1,6 @@
 import { nonce } from './options.js';
-import { createBlob, baseUrl, noop } from './common.js';
+import { createBlob, baseUrl } from './common.js';
+import noop from './noop.js';
 
 let err;
 window.addEventListener('error', _err => err = _err);
